tests/eventscripts: Be more defensive about removing the var directory
Could add -r but that might avoid error.  The only time it doesn't
exist is the 1st time a top-level var directory is used.


(This used to be ctdb commit 047d97dc84a5943556efb31bb2856fd04835fffa)

+++ ctdb/tests/eventscripts/scripts/local.sh
@@ -55,7 +55,8 @@ EOF
 fi
 
 export EVENTSCRIPTS_TESTS_VAR_DIR="${TEST_VAR_DIR}/unit_eventscripts"
-if [ "$EVENTSCRIPTS_TESTS_VAR_DIR" != "/unit_eventscripts" ] ; then
+if [ -d "$EVENTSCRIPTS_TESTS_VAR_DIR" -a \
+    "$EVENTSCRIPTS_TESTS_VAR_DIR" != "/unit_eventscripts" ] ; then
     rm -r "$EVENTSCRIPTS_TESTS_VAR_DIR"
 fi
 mkdir -p "$EVENTSCRIPTS_TESTS_VAR_DIR"
